+++ test/reflection.jl
@@ -221,8 +221,8 @@ mutable struct TLayout
 end
 tlayout = TLayout(5,7,11)
 @test fieldnames(TLayout) == (:x, :y, :z) == Base.propertynames(tlayout)
-@test hasfield(tlayout, :y)
-@test !hasfield(tlayout, :a)
+@test hasfield(TLayout, :y)
+@test !hasfield(TLayout, :a)
 @test hasproperty(tlayout, :x)
 @test !hasproperty(tlayout, :p)
 @test [(fieldoffset(TLayout,i), fieldname(TLayout,i), fieldtype(TLayout,i)) for i = 1:fieldcount(TLayout)] ==